requirements.txt
beautifulsoup4
enum34;python_version<"3.4"
websocket-client

aiopg==1.0.0
peewee==3.13.2
peewee-async==0.6.3a0
peewee-migrate==1.1.6
psycopg2-binary==2.8.4
langdetect==1.0.8